+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-06 09:10:58
+
+### Changed Files
+- `src/features/expenses/leads/LeadPurchaseDashboard.tsx`
+
+### Statistics
+```
+ 1 file changed, 2 insertions(+), 3 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-06 09:07:22
 
 ### Changed Files